+++ internal/utils/general.go
@@ -1,5 +1,7 @@
 package utils
 
+import "github.com/elliotchance/pie/pie"
+
 func Contains(s []string, e string) bool {
 	for _, a := range s {
 		if a == e {
@@ -13,4 +15,12 @@ func AppendUnique(element *[]string, elementToAppend string) {
 	if !Contains((*element), elementToAppend) {
 		(*element) = append((*element), elementToAppend)
 	}
+}
+
+func FilterList(list []string, valuesToFilter []string) []string {
+	pieList := pie.Strings(list)
+	newList := pieList.Filter(func(listItem string) bool {
+		return !Contains(valuesToFilter, listItem)
+	})
+	return newList
 }